Add task to pipeline to install kubectl
Kubectl is the CLI tool for interacting with Kubernetes clusters.
This commit adds a task to the continuous deployment Azure Pipeline
that installs kubectl onto the virtual machine running the pipeline.
We use environment variables to set the desired version of the tool
so that these can be checked/updated in one, findable place within
the file.