Install JupyterHub helm chart onto Kubernetes cluster
This commit adds a task to the continuous deployment azure pipeline that will upgrade/install the JupyterHub helm chart onto the Kubernetes cluster. We use the same value for the Kubernetes namespace and Helm release name for the sake of ease. The version of the helm chart we wish to install is defined as an environment variable so that it can be easily upgraded.