[DOCS] Add comments to Azure pipeline
The version of the HELM CHART is not a direct translation to the
version of the JUPYTERHUB code base we are installing. Version
0.8.2 of the HELM CHART corresponds to version 0.9.6 of the
JUPYTERHUB code base. A comment reflecting this and linking to the
z2jh-k8s tag documentation has been added to the Azure Pipeline
above the environment variables where we set the Helm Chart version